4.29 백준 7-4 내용변경 . 7-5,6은 풀이 .

+++ Algorithm/chapter7/Q4.java
@@ -6,13 +6,11 @@ public class Q4 {
 	public static void main(String[] args) {
 		Scanner sc = new Scanner(System.in);
 
-		int A = sc.nextInt();
-		int B = sc.nextInt();
-		int V = sc.nextInt();
-
+		long A = sc.nextInt();
+		long B = sc.nextInt();
+		long V = sc.nextInt();
 		sc.close();
-
-		int cnt = (V - B) / (A - B);
+		long cnt = (V - B) / (A - B);
 
 		if ((V - B) % (A - B) != 0) {
 			cnt++;
